+++ src/se/mod.rs
@@ -77,31 +77,24 @@ mod content;
 mod element;
 mod key;
 pub(crate) mod simple_type;
-mod var;
 
-use self::var::{Map, Seq, Struct, Tuple};
-use crate::{
-    de::PRIMITIVE_PREFIX,
-    errors::serialize::DeError,
-    events::{BytesEnd, BytesStart, BytesText, Event},
-    writer::{Indentation, Writer},
-};
+use self::content::ContentSerializer;
+use self::element::ElementSerializer;
+use crate::errors::serialize::DeError;
+use crate::writer::Indentation;
 use serde::ser::{self, Serialize};
 use serde::serde_if_integer128;
-use std::io::Write;
+use std::fmt::Write;
 use std::str::from_utf8;
 
 /// Serialize struct into a `Write`r
-pub fn to_writer<W: Write, S: Serialize>(writer: W, value: &S) -> Result<(), DeError> {
-    let mut serializer = Serializer::new(writer);
-    value.serialize(&mut serializer)
+pub fn to_writer<W: Write, S: Serialize>(writer: W, value: &S) -> Result<W, DeError> {
+    value.serialize(Serializer::new(writer))
 }
 
 /// Serialize struct into a `String`
 pub fn to_string<S: Serialize>(value: &S) -> Result<String, DeError> {
-    let mut writer = Vec::new();
-    to_writer(&mut writer, value)?;
-    Ok(String::from_utf8(writer)?)
+    to_writer(String::new(), value)
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -109,7 +102,7 @@ pub fn to_string<S: Serialize>(value: &S) -> Result<String, DeError> {
 /// Defines which characters would be escaped in [`Text`] events and attribute
 /// values.
 ///
-/// [`Text`]: Event::Text
+/// [`Text`]: crate::events::Event::Text
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum QuoteLevel {
     /// Performs escaping, escape all characters that could have special meaning
@@ -150,6 +143,18 @@ pub enum QuoteLevel {
     /// `<`      | `&lt;`
     /// `&`      | `&amp;`
     Minimal,
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////
+
+/// Implements serialization method by forwarding it to the serializer created by
+/// the helper method [`Serializer::ser`].
+macro_rules! forward {
+    ($name:ident($ty:ty)) => {
+        fn $name(self, value: $ty) -> Result<Self::Ok, Self::Error> {
+            self.ser(&concat!("`", stringify!($ty), "`"))?.$name(value)
+        }
+    };
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -273,9 +278,9 @@ impl<'i> Indent<'i> {
 
 /// A Serializer
 pub struct Serializer<'r, W: Write> {
-    writer: Writer<W>,
+    ser: ContentSerializer<'r, W>,
     /// Name of the root tag. If not specified, deduced from the structure name
-    root_tag: Option<&'r str>,
+    root_tag: Option<XmlName<'r>>,
 }
 
 impl<'r, W: Write> Serializer<'r, W> {
@@ -285,10 +290,19 @@ impl<'r, W: Write> Serializer<'r, W> {
     /// and newtype structs) will end up to an error. Use `with_root` to create
     /// serializer with explicitly defined root element name
     pub fn new(writer: W) -> Self {
-        Self::with_root(Writer::new(writer), None)
+        Self {
+            ser: ContentSerializer {
+                writer,
+                level: QuoteLevel::Full,
+                indent: Indent::None,
+                write_indent: false,
+            },
+            root_tag: None,
+        }
     }
 
-    /// Creates a new `Serializer` that uses specified root tag name
+    /// Creates a new `Serializer` that uses specified root tag name. `name` should
+    /// be valid [XML name], otherwise error is returned.
     ///
     /// # Examples
     ///
@@ -298,14 +312,13 @@ impl<'r, W: Write> Serializer<'r, W> {
     /// # use pretty_assertions::assert_eq;
     /// # use serde::Serialize;
     /// # use quick_xml::se::Serializer;
-    /// use quick_xml::writer::Writer;
     ///
-    /// let mut buffer = Vec::new();
-    /// let mut writer = Writer::new_with_indent(&mut buffer, b' ', 2);
-    /// let mut ser = Serializer::with_root(writer, Some("root"));
+    /// let ser = Serializer::with_root(String::new(), Some("root")).unwrap();
     ///
-    /// "node".serialize(&mut ser).unwrap();
-    /// assert_eq!(String::from_utf8(buffer).unwrap(), "node");
+    /// assert_eq!(
+    ///     "node".serialize(ser).unwrap(),
+    ///     "<root>node</root>"
+    /// );
     /// ```
     ///
     /// When serializing a struct, newtype struct, unit struct or tuple `root_tag`
@@ -314,8 +327,7 @@ impl<'r, W: Write> Serializer<'r, W> {
     /// ```
     /// # use pretty_assertions::assert_eq;
     /// # use serde::Serialize;
-    /// use quick_xml::se::Serializer;
-    /// use quick_xml::writer::Writer;
+    /// # use quick_xml::se::Serializer;
     ///
     /// #[derive(Debug, PartialEq, Serialize)]
     /// struct Struct {
@@ -323,144 +335,107 @@ impl<'r, W: Write> Serializer<'r, W> {
     ///     answer: u32,
     /// }
     ///
-    /// let mut buffer = Vec::new();
-    /// let mut writer = Writer::new_with_indent(&mut buffer, b' ', 2);
-    /// let mut ser = Serializer::with_root(writer, Some("root"));
+    /// let ser = Serializer::with_root(String::new(), Some("root")).unwrap();
     ///
-    /// Struct {
+    /// let data = Struct {
     ///     question: "The Ultimate Question of Life, the Universe, and Everything".into(),
     ///     answer: 42,
-    /// }.serialize(&mut ser).unwrap();
+    /// };
+    ///
     /// assert_eq!(
-    ///     String::from_utf8(buffer.clone()).unwrap(),
-    ///     r#"<root question="The Ultimate Question of Life, the Universe, and Everything" answer="42"/>"#
+    ///     data.serialize(ser).unwrap(),
+    ///     "<root>\
+    ///         <question>The Ultimate Question of Life, the Universe, and Everything</question>\
+    ///         <answer>42</answer>\
+    ///      </root>"
     /// );
     /// ```
-    pub fn with_root(writer: Writer<W>, root_tag: Option<&'r str>) -> Self {
-        Self { writer, root_tag }
+    ///
+    /// [XML name]: https://www.w3.org/TR/REC-xml/#NT-Name
+    pub fn with_root(writer: W, root_tag: Option<&'r str>) -> Result<Self, DeError> {
+        Ok(Self {
+            ser: ContentSerializer {
+                writer,
+                level: QuoteLevel::Full,
+                indent: Indent::None,
+                write_indent: false,
+            },
+            root_tag: root_tag.map(|tag| XmlName::try_from(tag)).transpose()?,
+        })
     }
 
-    fn write_primitive<P: std::fmt::Display>(
-        &mut self,
-        value: P,
-        escaped: bool,
-    ) -> Result<(), DeError> {
-        let value = value.to_string();
-        let event = if escaped {
-            BytesText::from_escaped(&value)
+    /// Configure indent for a serializer
+    pub fn indent(&mut self, indent_char: char, indent_size: usize) -> &mut Self {
+        self.ser.indent = Indent::Owned(Indentation::new(indent_char as u8, indent_size));
+        self
+    }
+
+    /// Creates actual serializer or returns an error if root tag is not defined.
+    /// In that case `err` contains the name of type that cannot be serialized.
+    fn ser(self, err: &str) -> Result<ElementSerializer<'r, W>, DeError> {
+        if let Some(key) = self.root_tag {
+            Ok(ElementSerializer { ser: self.ser, key })
         } else {
-            BytesText::new(&value)
-        };
-        self.writer.write_event(Event::Text(event))?;
-        Ok(())
+            Err(DeError::Unsupported(
+                format!("cannot serialize {} without defined root tag", err).into(),
+            ))
+        }
     }
 
-    /// Writes self-closed tag `<tag_name/>` into inner writer
-    fn write_self_closed(&mut self, tag_name: &str) -> Result<(), DeError> {
-        self.writer
-            .write_event(Event::Empty(BytesStart::new(tag_name)))?;
-        Ok(())
-    }
-
-    /// Writes a serialized `value` surrounded by `<tag_name>...</tag_name>`
-    fn write_paired<T: ?Sized + Serialize>(
-        &mut self,
-        tag_name: &str,
-        value: &T,
-    ) -> Result<(), DeError> {
-        self.writer
-            .write_event(Event::Start(BytesStart::new(tag_name)))?;
-        value.serialize(&mut *self)?;
-        self.writer
-            .write_event(Event::End(BytesEnd::new(tag_name)))?;
-        Ok(())
+    /// Creates actual serializer using root tag or a specified `key` if root tag
+    /// is not defined. Returns an error if root tag is not defined and a `key`
+    /// does not conform [XML rules](XmlName::try_from) for names.
+    fn ser_name(self, key: &'static str) -> Result<ElementSerializer<'r, W>, DeError> {
+        Ok(ElementSerializer {
+            ser: self.ser,
+            key: match self.root_tag {
+                Some(key) => key,
+                None => XmlName::try_from(key)?,
+            },
+        })
     }
 }
 
-impl<'r, 'w, W: Write> ser::Serializer for &'w mut Serializer<'r, W> {
-    type Ok = ();
+impl<'r, W: Write> ser::Serializer for Serializer<'r, W> {
+    type Ok = W;
     type Error = DeError;
 
-    type SerializeSeq = Seq<'r, 'w, W>;
-    type SerializeTuple = Tuple<'r, 'w, W>;
-    type SerializeTupleStruct = Tuple<'r, 'w, W>;
-    type SerializeTupleVariant = Tuple<'r, 'w, W>;
-    type SerializeMap = Map<'r, 'w, W>;
-    type SerializeStruct = Struct<'r, 'w, W>;
-    type SerializeStructVariant = Struct<'r, 'w, W>;
+    type SerializeSeq = <ElementSerializer<'r, W> as ser::Serializer>::SerializeSeq;
+    type SerializeTuple = <ElementSerializer<'r, W> as ser::Serializer>::SerializeTuple;
+    type SerializeTupleStruct = <ElementSerializer<'r, W> as ser::Serializer>::SerializeTupleStruct;
+    type SerializeTupleVariant =
+        <ElementSerializer<'r, W> as ser::Serializer>::SerializeTupleVariant;
+    type SerializeMap = <ElementSerializer<'r, W> as ser::Serializer>::SerializeMap;
+    type SerializeStruct = <ElementSerializer<'r, W> as ser::Serializer>::SerializeStruct;
+    type SerializeStructVariant =
+        <ElementSerializer<'r, W> as ser::Serializer>::SerializeStructVariant;
 
-    fn serialize_bool(self, v: bool) -> Result<Self::Ok, DeError> {
-        self.write_primitive(if v { "true" } else { "false" }, true)
-    }
+    forward!(serialize_bool(bool));
 
-    fn serialize_i8(self, v: i8) -> Result<Self::Ok, DeError> {
-        self.write_primitive(v, true)
-    }
+    forward!(serialize_i8(i8));
+    forward!(serialize_i16(i16));
+    forward!(serialize_i32(i32));
+    forward!(serialize_i64(i64));
 
-    fn serialize_i16(self, v: i16) -> Result<Self::Ok, DeError> {
-        self.write_primitive(v, true)
-    }
-
-    fn serialize_i32(self, v: i32) -> Result<Self::Ok, DeError> {
-        self.write_primitive(v, true)
-    }
-
-    fn serialize_i64(self, v: i64) -> Result<Self::Ok, DeError> {
-        self.write_primitive(v, true)
-    }
-
-    fn serialize_u8(self, v: u8) -> Result<Self::Ok, DeError> {
-        self.write_primitive(v, true)
-    }
-
-    fn serialize_u16(self, v: u16) -> Result<Self::Ok, DeError> {
-        self.write_primitive(v, true)
-    }
-
-    fn serialize_u32(self, v: u32) -> Result<Self::Ok, DeError> {
-        self.write_primitive(v, true)
-    }
-
-    fn serialize_u64(self, v: u64) -> Result<Self::Ok, DeError> {
-        self.write_primitive(v, true)
-    }
+    forward!(serialize_u8(u8));
+    forward!(serialize_u16(u16));
+    forward!(serialize_u32(u32));
+    forward!(serialize_u64(u64));
 
     serde_if_integer128! {
-        fn serialize_i128(self, v: i128) -> Result<Self::Ok, DeError> {
-            self.write_primitive(v, true)
-        }
-
-        fn serialize_u128(self, v: u128) -> Result<Self::Ok, DeError> {
-            self.write_primitive(v, true)
-        }
+        forward!(serialize_i128(i128));
+        forward!(serialize_u128(u128));
     }
 
-    fn serialize_f32(self, v: f32) -> Result<Self::Ok, DeError> {
-        self.write_primitive(v, true)
-    }
+    forward!(serialize_f32(f32));
+    forward!(serialize_f64(f64));
 
-    fn serialize_f64(self, v: f64) -> Result<Self::Ok, DeError> {
-        self.write_primitive(v, true)
-    }
-
-    fn serialize_char(self, v: char) -> Result<Self::Ok, DeError> {
-        self.write_primitive(v, false)
-    }
-
-    fn serialize_str(self, value: &str) -> Result<Self::Ok, DeError> {
-        self.write_primitive(value, false)
-    }
-
-    fn serialize_bytes(self, _value: &[u8]) -> Result<Self::Ok, DeError> {
-        // TODO: I imagine you'd want to use base64 here.
-        // Not sure how to roundtrip effectively though...
-        Err(DeError::Unsupported(
-            "`serialize_bytes` not supported yet".into(),
-        ))
-    }
+    forward!(serialize_char(char));
+    forward!(serialize_str(&str));
+    forward!(serialize_bytes(&[u8]));
 
     fn serialize_none(self) -> Result<Self::Ok, DeError> {
-        Ok(())
+        Ok(self.ser.writer)
     }
 
     fn serialize_some<T: ?Sized + Serialize>(self, value: &T) -> Result<Self::Ok, DeError> {
@@ -468,25 +443,21 @@ impl<'r, 'w, W: Write> ser::Serializer for &'w mut Serializer<'r, W> {
     }
 
     fn serialize_unit(self) -> Result<Self::Ok, DeError> {
-        self.serialize_none()
+        self.ser("`()`")?.serialize_unit()
     }
 
     fn serialize_unit_struct(self, name: &'static str) -> Result<Self::Ok, DeError> {
-        self.write_self_closed(self.root_tag.unwrap_or(name))
+        self.ser_name(name)?.serialize_unit_struct(name)
     }
 
     fn serialize_unit_variant(
         self,
-        _name: &'static str,
-        _variant_index: u32,
+        name: &'static str,
+        variant_index: u32,
         variant: &'static str,
     ) -> Result<Self::Ok, DeError> {
-        if variant.starts_with(PRIMITIVE_PREFIX) {
-            let variant = variant.split_at(PRIMITIVE_PREFIX.len()).1;
-            self.write_primitive(variant, false)
-        } else {
-            self.write_self_closed(variant)
-        }
+        self.ser_name(name)?
+            .serialize_unit_variant(name, variant_index, variant)
     }
 
     fn serialize_newtype_struct<T: ?Sized + Serialize>(
@@ -494,85 +465,67 @@ impl<'r, 'w, W: Write> ser::Serializer for &'w mut Serializer<'r, W> {
         name: &'static str,
         value: &T,
     ) -> Result<Self::Ok, DeError> {
-        self.write_paired(self.root_tag.unwrap_or(name), value)
+        self.ser_name(name)?.serialize_newtype_struct(name, value)
     }
 
     fn serialize_newtype_variant<T: ?Sized + Serialize>(
         self,
-        _name: &'static str,
-        _variant_index: u32,
+        name: &'static str,
+        variant_index: u32,
         variant: &'static str,
         value: &T,
     ) -> Result<Self::Ok, DeError> {
-        // Flatten structs in enums are serialized as newtype struct variant + map.
-        // As serialize_map should write `root_tag` for ordinal maps (because it's
-        // only way for maps), and for enums this method already written a tag name
-        // (`variant`), we need to clear root tag before writing content and restore
-        // it after
-        let root = self.root_tag.take();
-        let result = self.write_paired(variant, value);
-        self.root_tag = root;
-        result
+        self.ser_name(name)?
+            .serialize_newtype_variant(name, variant_index, variant, value)
     }
 
-    fn serialize_seq(self, _len: Option<usize>) -> Result<Self::SerializeSeq, DeError> {
-        Ok(Seq::new(self))
+    fn serialize_seq(self, len: Option<usize>) -> Result<Self::SerializeSeq, DeError> {
+        self.ser("sequence")?.serialize_seq(len)
     }
 
-    fn serialize_tuple(self, _len: usize) -> Result<Self::SerializeTuple, DeError> {
-        let tag = match self.root_tag {
-            Some(tag) => tag,
-            None => {
-                return Err(DeError::Custom(
-                    "cannot serialize unnamed tuple without defined root tag".into(),
-                ))
-            }
-        };
-        Ok(Tuple::new(self, tag))
+    fn serialize_tuple(self, len: usize) -> Result<Self::SerializeTuple, DeError> {
+        self.ser("unnamed tuple")?.serialize_tuple(len)
     }
 
     fn serialize_tuple_struct(
         self,
         name: &'static str,
-        _len: usize,
+        len: usize,
     ) -> Result<Self::SerializeTupleStruct, DeError> {
-        Ok(Tuple::new(self, self.root_tag.unwrap_or(name)))
+        self.ser_name(name)?.serialize_tuple_struct(name, len)
     }
 
     fn serialize_tuple_variant(
         self,
-        _name: &'static str,
-        _variant_index: u32,
+        name: &'static str,
+        variant_index: u32,
         variant: &'static str,
-        _len: usize,
+        len: usize,
     ) -> Result<Self::SerializeTupleVariant, DeError> {
-        Ok(Tuple::new(self, variant))
+        self.ser_name(name)?
+            .serialize_tuple_variant(name, variant_index, variant, len)
     }
 
-    fn serialize_map(self, _len: Option<usize>) -> Result<Self::SerializeMap, DeError> {
-        if let Some(tag) = self.root_tag {
-            // TODO: Write self-closed tag if map is empty
-            self.writer
-                .write_event(Event::Start(BytesStart::new(tag)))?;
-        }
-        Ok(Map::new(self))
+    fn serialize_map(self, len: Option<usize>) -> Result<Self::SerializeMap, DeError> {
+        self.ser("map")?.serialize_map(len)
     }
 
     fn serialize_struct(
         self,
         name: &'static str,
-        _len: usize,
+        len: usize,
     ) -> Result<Self::SerializeStruct, DeError> {
-        Ok(Struct::new(self, self.root_tag.unwrap_or(name)))
+        self.ser_name(name)?.serialize_struct(name, len)
     }
 
     fn serialize_struct_variant(
         self,
-        _name: &'static str,
-        _variant_index: u32,
+        name: &'static str,
+        variant_index: u32,
         variant: &'static str,
-        _len: usize,
+        len: usize,
     ) -> Result<Self::SerializeStructVariant, DeError> {
-        Ok(Struct::new(self, variant))
+        self.ser_name(name)?
+            .serialize_struct_variant(name, variant_index, variant, len)
     }
 }